requirements.txt
flask==3.1.*
flask-cors==5.*
gunicorn==23.*
pyyaml==6.*
PyJWT==2.*
bcrypt==4.*
